requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-debug-toolbar==4.3.0
django-filter==24.2
djangorestframework==3.15.1
docker==7.0.0
docutils==0.20.1
drf-spectacular==0.27.1
flake8==7.0.0
flake8-docstrings==1.7.0
gunicorn==22.0.0
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mccabe==0.7.0
meta==1.0.2
numpy==1.26.4
packaging==24.0
pillow==10.2.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pydocstyle==6.3.0
pyflakes==3.2.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
sentry-sdk==2.1.1
snowballstemmer==2.2.0
sqlparse==0.4.4
style==1.1.0
update==0.0.1
uritemplate==4.1.1
urllib3==2.2.1
